msm: kgsl: Change ioctl memory sizes to size_t
Change all memory sizes in ioctl structs to size_t to maintain
consistency. Fix print messages to account for size_t. Also change
padding to unsigned long in memory related ioctls as new fields
are likely to be pointers or sizes.

+++ drivers/gpu/msm/kgsl_cffdump.c
@@ -412,7 +412,7 @@ void kgsl_cffdump_user_event(struct kgsl_device *device,
 
 void kgsl_cffdump_syncmem(struct kgsl_device *device,
 			  struct kgsl_memdesc *memdesc, uint gpuaddr,
-			  uint sizebytes, bool clean_cache)
+			  size_t sizebytes, bool clean_cache)
 {
 	const void *src;
